refactor: artistId 타입 Long으로 변경
change artistId type from Integer to Long across related classes

+++ backend/src/main/java/com/a404/duckonback/dto/RoomDTO.java
@@ -16,7 +16,7 @@ public class RoomDTO {
     private String imgUrl;
     private LocalDateTime createdAt;
     private String creatorId;
-    private int artistId;
+    private Long artistId;
 
     public static RoomDTO fromEntity(Room room) {
         return RoomDTO.builder()
